// src/api/__generated__/wallet/model/assetListResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */
import type { BalanceItemResponse } from './balanceItemResponse';

/**
 * 资产列表响应
 */
export interface AssetListResponse {
  /** 资产余额列表 */
  balances?: BalanceItemResponse[];
}