// src/api/__generated__/wallet/model/balanceItemResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */

/**
 * 资产余额
 */
export interface BalanceItemResponse {
  /** 资产编码：STORY、USDC */
  assetCode?: string;
  /** 可用余额 */
  availableBalance?: string;
  /** 冻结余额 */
  frozenBalance?: string;
  /** 资产精度 */
  decimals?: string;
}